prepare for search module
